requirements.txt
python==3.9
pytorch==1.13.0
torchvision
pytorch-cuda==11.6
iopath
jupyter
scikit-image
matplotlib
imageio
plotly
opencv-python
pillow
fvcore
black
usort
flake8
flake8-bugbear
flake8-comprehensions
